:truck: 지상 이동 로직 GroundMovementComponent 로 이동
