+++ Runelight-Smash-2021/Assets/Scripts/Component/GroundMovementComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d73a4b183b76d7a438da8d332bbc668f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 150
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
